requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
boto3==1.34.15
botocore==1.34.15
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
distro==1.9.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
openai==1.7.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.32.0.post1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
